requirements.txt
Django==5.0.1
django-unfold==0.20.0
django-htmx==1.18.0
django-celery-beat==2.6.0
django-celery-results==2.5.1
celery==5.3.6
redis==5.0.1
django-crispy-forms==2.1
crispy-tailwind==0.7.0
django-ckeditor==6.7.0
django-taggit==5.0.1
django-allauth==0.60.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
whitenoise==6.6.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
python-decouple==3.8
pillow==10.2.0
gunicorn==21.2.0
requests==2.31.0
django-import-export==3.3.5
